chore: move instance launch to separate file
